move common test fixtures into higher-scoped conftest.py. add `fixtures` test module for common fixture tests.
